Fixed item_db/item_db2 SQL structure Harmonized NPC events implementation
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3762 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ sql-files/item_db.sql
@@ -15,7 +15,7 @@ CREATE TABLE `item_db` (
   `defence` tinyint(3) unsigned default NULL,
   `range` tinyint(2) unsigned default NULL,
   `slots` tinyint(2) unsigned default NULL,
-  `equip_jobs` mediumint(8) unsigned default NULL,
+  `equip_jobs` int(12) unsigned default NULL,
   `equip_upper` tinyint(8) unsigned default NULL,
   `equip_genders` tinyint(2) unsigned default NULL,
   `equip_locations` smallint(4) unsigned default NULL,
